# roachpb: micro-optimize Value.SetProto