```
name             old time/op    new time/op    delta
ValueSetProto-8     103ns ± 1%      73ns ± 2%  -29.60%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
ValueSetProto-8      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=10+10)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/9861)
<!-- Reviewable:end -->