@tamird 2016-10-10
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 1 of 1 files at r1.
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[roachpb/data.go, line 365 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861#-KTkzlDQjqXJ7Zd9H_rC:-KTkzlDQjqXJ7Zd9H_rD:b-3r84m0) ([raw file](https://github.com/cockroachdb/cockroach/blob/6da548418f67cb14e82708a37479c5bac8862a47/roachpb/data.go#L365)):*
> ```Go
> 		size := m.Size()
> 		v.RawBytes = make([]byte, headerSize+size)
> 		protoutil.Interceptor(msg)
> ```

Isn't this a bit fragile? It might be better for `MarshalTo` to be a static function in `protoutil`.

note that `protoutil.Interceptor` is never called outside `protoutil` currently.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861#-:-KTl-SH1hTuhcXu-qAdM:bnfp4nl)*
<!-- Sent from Reviewable.io -->