@petermattis 2016-10-11
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[roachpb/data.go, line 365 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861#-KTkzlDQjqXJ7Zd9H_rC:-KTl8UfRaUuNBCkKae_K:b-tu7l17) ([raw file](https://github.com/cockroachdb/cockroach/blob/6da548418f67cb14e82708a37479c5bac8862a47/roachpb/data.go#L365)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary>
> Isn't this a bit fragile? It might be better for `MarshalTo` to be a static function in `protoutil`.
> 
> note that `protoutil.Interceptor` is never called outside `protoutil` currently.
</details>
I had that same thought at dinner and was wondering if you were going to call me on it. Unfortunately, using a `protoutil.MarshalTo` is significantly slower than this current PR, eliminating all of the gains and then some (`112 ns/op`). 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861)*
<!-- Sent from Reviewable.io -->