@tamird 2016-10-11
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[roachpb/data.go, line 365 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861#-KTkzlDQjqXJ7Zd9H_rC:-KTl9BpS4AYhPskJqOPm:b-c5x9nq) ([raw file](https://github.com/cockroachdb/cockroach/blob/6da548418f67cb14e82708a37479c5bac8862a47/roachpb/data.go#L365)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary>
> I had that same thought at dinner and was wondering if you were going to call me on it. Unfortunately, using a `protoutil.MarshalTo` is significantly slower than this current PR, eliminating all of the gains and then some (`112 ns/op`). 
</details>
Really? That's surprising. `MarshalTo` is always dynamically dispatched here (and thus never inlined), so...the function call overhead of `protoutil.MarshalTo` accounts for all the speed difference?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861)*
<!-- Sent from Reviewable.io -->