@petermattis 2016-10-11
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[roachpb/data.go, line 365 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861#-KTkzlDQjqXJ7Zd9H_rC:-KTlC35ZncQYJUaXSC8J:bubzdxo) ([raw file](https://github.com/cockroachdb/cockroach/blob/d6ca7c95c1630e1658665cc43d6e26d129f1a055/roachpb/data.go#L365)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary>
> Really? That's surprising. `MarshalTo` is always dynamically dispatched here (and thus never inlined), so...the function call overhead of `protoutil.MarshalTo` accounts for all the speed difference?
</details>
I agree it is surprising. I'm not sure exactly what the difference is, but it is very repeatable. Cpu profiling shows different functions showing up, almost as if using `protoutil.MarshalTo` affected inlining significantly. That seems weird/unexpected.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861)*
<!-- Sent from Reviewable.io -->