@tamird 2016-10-11
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[roachpb/data.go, line 365 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861#-KTkzlDQjqXJ7Zd9H_rC:-KTlCc6B8XD4W7oY5Zos:b-djbv06) ([raw file](https://github.com/cockroachdb/cockroach/blob/d6ca7c95c1630e1658665cc43d6e26d129f1a055/roachpb/data.go#L365)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary>
> I agree it is surprising. I'm not sure exactly what the difference is, but it is very repeatable. Cpu profiling shows different functions showing up, almost as if using `protoutil.MarshalTo` affected inlining significantly. That seems weird/unexpected.
</details>
Sigh. LGTM with a comment.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861)*
<!-- Sent from Reviewable.io -->