@petermattis 2016-10-11
Review status: 0 of 1 files reviewed at latest revision, 1 unresolved discussion.

---

*[roachpb/data.go, line 365 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861#-KTkzlDQjqXJ7Zd9H_rC:-KTlEyz86U4DPNQ9ND2_:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/d6ca7c95c1630e1658665cc43d6e26d129f1a055/roachpb/data.go#L365)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary>
> Sigh. LGTM with a comment.
</details>
Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/9861)*
<!-- Sent from Reviewable.io -->